requirements.txt
certifi==2025.11.12
python-dotenv==1.2.2
charset-normalizer==3.4.4
greenlet==3.2.4
idna==3.11
playwright==1.56.0
pyee==13.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
